Struct probabilistic_collections::bloom::BSBloomFilter

pub struct BSBloomFilter<T, B = SipHasherBuilder> { /* fields omitted */ }

A space-efficient probabilistic data structure for data deduplication in streams.

This particular implementation uses Biased Sampling to determine whether data is distinct. Past work for data deduplication include Stable Bloom Filters, but it suffers from a high false negative rate and slow convergence rate.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let mut filter = BSBloomFilter::<String>::new(10, 0.01);

assert!(!filter.contains("foo"));
filter.insert("foo");
assert!(filter.contains("foo"));

filter.clear();
assert!(!filter.contains("foo"));

assert_eq!(filter.len(), 70);
assert_eq!(filter.bit_count(), 10);
assert_eq!(filter.hasher_count(), 7);

Methods

impl<T> BSBloomFilter<T>

pub fn new(bit_count: usize, fpp: f64) -> Self

Constructs a new, empty BSBloomFilter with bit_count bits per filter, and a false positive probability of fpp.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let filter = BSBloomFilter::<String>::new(10, 0.01);

impl<T, B> BSBloomFilter<T, B> where
B: BuildHasher,

pub fn with_hashers(bit_count: usize, fpp: f64, hash_builders: [B; 2]) -> Self

Constructs a new, empty BSBloomFilter with bit_count bits per filter, a false positive probability of fpp, and two hash builders for double hashing.

Examples

use probabilistic_collections::bloom::BSBloomFilter;
use probabilistic_collections::SipHasherBuilder;

let filter = BSBloomFilter::<String>::with_hashers(
    10,
    0.01,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

pub fn insert<U: ?Sized>(&mut self, item: &U) where
T: Borrow<U>,
U: Hash,

Inserts an element into the bloom filter and returns if it is distinct

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let mut filter = BSBloomFilter::<String>::new(10, 0.01);

filter.insert("foo");

pub fn contains<U: ?Sized>(&self, item: &U) -> bool where
T: Borrow<U>,
U: Hash,

Checks if an element is possibly in the bloom filter.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let mut filter = BSBloomFilter::<String>::new(10, 0.01);

assert!(!filter.contains("foo"));
filter.insert("foo");
assert!(filter.contains("foo"));

pub fn len(&self) -> usize

Returns the number of bits in the bloom filter.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let filter = BSBloomFilter::<String>::new(10, 0.01);

assert_eq!(filter.len(), 70);

pub fn is_empty(&self) -> bool

Returns true if the bloom filter is empty.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let filter = BSBloomFilter::<String>::new(10, 0.01);

assert!(!filter.is_empty());

pub fn bit_count(&self) -> usize

Returns the number of bits in each partition in the bloom filter.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let filter = BSBloomFilter::<String>::new(10, 0.01);

assert_eq!(filter.bit_count(), 10);

pub fn hasher_count(&self) -> usize

Returns the number of hash functions used by the bloom filter.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let filter = BSBloomFilter::<String>::new(10, 0.01);

assert_eq!(filter.hasher_count(), 7);

pub fn clear(&mut self)

Clears the bloom filter, removing all elements.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let mut filter = BSBloomFilter::<String>::new(10, 0.01);

filter.insert("foo");
filter.clear();

assert!(!filter.contains("foo"));

pub fn count_ones(&self) -> usize

Returns the number of set bits in the bloom filter.

Examples

use probabilistic_collections::bloom::BSBloomFilter;
use probabilistic_collections::SipHasherBuilder;

let mut filter = BSBloomFilter::<String>::with_hashers(
    10,
    0.01,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);
filter.insert("foo");

assert_eq!(filter.count_ones(), 7);

pub fn count_zeros(&self) -> usize

Returns the number of unset bits in the bloom filter.

Examples

use probabilistic_collections::bloom::BSBloomFilter;
use probabilistic_collections::SipHasherBuilder;

let mut filter = BSBloomFilter::<String>::with_hashers(
    10,
    0.01,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);
filter.insert("foo");

assert_eq!(filter.count_zeros(), 63);

pub fn hashers(&self) -> &[B; 2]

Returns a reference to the bloom filter's hasher builders.

Examples

use probabilistic_collections::bloom::BSBloomFilter;

let filter = BSBloomFilter::<String>::new(10, 0.01);
let hashers = filter.hashers();