Module probabilistic_collections::count_min_sketch

Space-efficient probabilistic data structure for estimating the number of item occurrences.

Structs

CountMeanStrategy
A count strategy that uses the mean value to estimate the number of item occurrences. This strategy performs well when there are item removals or if some values are negative.

CountMedianBiasStrategy
A count strategy that uses the median value after substracting a bias from all values to mitigate the effects of overestimation. This strategy may also overestimate, so we take the minimum of the CountMedianBiasStrategy and the CountMinStrategy to get a more accurate estimate.

CountMinSketch
A space-efficient probabilistic data structure that serves as a frequency table of events in a stream of data.

CountMinStrategy
A count strategy that uses the minimum value to estimate the number of item occurrences. This strategy is suspectible to overestimation, but never underestimation.

ItemValueIter
An iterator that yields values corresponding to an item in the count-min sketch.

Traits

CountStrategy
Trait for types that have the logic for estimating the number of item occurrences.