Trait probabilistic_collections::count_min_sketch::CountStrategy

pub trait CountStrategy {
    fn get_estimate(
        items: i64, 
        rows: usize, 
        cols: usize, 
        iter: ItemValueIter
    ) -> i64;
}

Trait for types that have the logic for estimating the number of item occurrences.

Required methods

fn get_estimate(
items: i64,
rows: usize,
cols: usize,
iter: ItemValueIter
) -> i64

Returns the estimated number of item occurrences given the number of items in the CountMinSketch, the rows in the grid, the columns in the grid, and the values corresponding to the item in the grid

Loading content...