Module probabilistic_collections::cuckoo

Space-efficient probabilistic data structure for approximate membership queries in a set with the ability to remove items.

Structs

CuckooFilter
A space-efficient probabilistic data structure to test for membership in a set. Cuckoo filters also provide the flexibility to remove items.

ScalableCuckooFilter
A growable, space-efficient probabilistic data structure to test for membership in a set. Scalable cuckoo filters also provide the flexibility to remove items.