Struct probabilistic_collections::cuckoo::CuckooFilter

pub struct CuckooFilter<T, B = SipHasherBuilder> { /* fields omitted */ }

A space-efficient probabilistic data structure to test for membership in a set. Cuckoo filters also provide the flexibility to remove items.

A cuckoo filter is based on cuckoo hashing and is essentially a cuckoo hash table storing each keys' fingerprint. Cuckoo filters can be highly compact and serve as an improvement over variations of tradition Bloom filters that support deletion (E.G. counting Bloom filters).

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let mut filter = CuckooFilter::<String>::new(100);

assert!(!filter.contains("foo"));
filter.insert("foo");
assert!(filter.contains("foo"));

filter.remove("foo");
assert!(!filter.contains("foo"));

assert_eq!(filter.len(), 0);
assert_eq!(filter.capacity(), 128);
assert_eq!(filter.bucket_len(), 32);
assert_eq!(filter.fingerprint_bit_count(), 8);

Methods

impl<T> CuckooFilter<T>

pub fn new(item_count: usize) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count. By default, the cuckoo filter will have 8 bits per item fingerprint, 4 entries per index, and a maximum of 512 item displacements before terminating the insertion process. The cuckoo filter will have an estimated maximum false positive probability of 3%.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);

pub fn from_parameters(
item_count: usize,
fingerprint_bit_count: usize,
entries_per_index: usize
) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count, a fingerprint bit count of fingerprint_bit_count, entries_per_index entries per index, and a maximum of 512 item displacements before terminating the insertion process. This method provides no guarantees on the false positive probability of the cuckoo filter.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0, if fingerprint_bit_count less than 1 or greater than 64, or if entries_per_index is 0.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::from_parameters(100, 16, 8);

pub fn from_entries_per_index(
item_count: usize,
fpp: f64,
entries_per_index: usize
) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count, an estimated maximum false positive probability of fpp, entries_per_index entries per index, and a maximum of 512 item displacements before terminating the insertion process.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0 or if entries_per_index is 0.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::from_entries_per_index(100, 0.01, 4);

pub fn from_fingerprint_bit_count(
item_count: usize,
fpp: f64,
fingerprint_bit_count: usize
) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count, an estimated maximum false positive probability of fpp, a fingerprint bit count of fingerprint_bit_count, and a maximum of 512 item displacements before terminating the insertion process.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0, if fingerprint_bit_count is less than 1 or greater than 64, or if it is impossible to achieve the given maximum false positive probability.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::from_fingerprint_bit_count(100, 0.01, 10);

impl<T, B> CuckooFilter<T, B> where
B: BuildHasher,

pub fn with_hashers(item_count: usize, hash_builders: [B; 2]) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count, and two hasher builders for double hashing. By default, the cuckoo filter will have 8 bits per item fingerprint, 4 entries per index, and a maximum of 512 item displacements before terminating the insertion process. The cuckoo filter will have an estimated maximum false positive probability of 3%.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;
use probabilistic_collections::SipHasherBuilder;

let filter = CuckooFilter::<String>::with_hashers(
    100,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

pub fn from_parameters_with_hashers(
item_count: usize,
fingerprint_bit_count: usize,
entries_per_index: usize,
hash_builders: [B; 2]
) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count, a fingerprint bit count of fingerprint_bit_count, entries_per_index entries per index, a maximum of 512 item displacements before terminating the insertion process, and two hasher builders for double hashing. This method provides no guarantees on the false positive probability of the cuckoo filter.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0, if fingerprint_bit_count less than 1 or greater than 64, or if entries_per_index is 0.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;
use probabilistic_collections::SipHasherBuilder;

let filter = CuckooFilter::<String>::from_parameters_with_hashers(
    100,
    16,
    8,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

pub fn from_entries_per_index_with_hashers(
item_count: usize,
fpp: f64,
entries_per_index: usize,
hash_builders: [B; 2]
) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count, an estimated maximum false positive probability of fpp, entries_per_index entries per index, a maximum of 512 item displacements before terminating the insertion process, and two hasher builders for doubling hashing.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0 or if entries_per_index is 0.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;
use probabilistic_collections::SipHasherBuilder;

let filter = CuckooFilter::<String>::from_entries_per_index_with_hashers(
    100,
    0.01,
    4,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

pub fn from_fingerprint_bit_count_with_hashers(
item_count: usize,
fpp: f64,
fingerprint_bit_count: usize,
hash_builders: [B; 2]
) -> Self

Constructs a new, empty CuckooFilter with an estimated max capacity of item_count, an estimated maximum false positive probability of fpp, a fingerprint bit count of fingerprint_bit_count, a maximum of 512 item displacements before terminating the insertion process, and two hasher builders for double hashing.

The length of each bucket will be rounded off to the next power of two.

Panics

Panics if item_count is 0, if fingerprint_bit_count is less than 1 or greater than 64, or if it is impossible to achieve the given maximum false positive probability.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;
use probabilistic_collections::SipHasherBuilder;

let filter = CuckooFilter::<String>::from_fingerprint_bit_count_with_hashers(
    100,
    0.01,
    10,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

pub fn insert<U: ?Sized>(&mut self, item: &U) where
T: Borrow<U>,
U: Hash,

Inserts an element into the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let mut filter = CuckooFilter::<String>::new(100);
filter.insert("foo");

pub fn remove<U: ?Sized>(&mut self, item: &U) where
T: Borrow<U>,
U: Hash,

Removes an element from the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let mut filter = CuckooFilter::<String>::new(100);

filter.insert("foo");
assert!(filter.contains("foo"));

filter.remove("foo");
assert!(!filter.contains("foo"));

pub fn contains<U: ?Sized>(&self, item: &U) -> bool where
T: Borrow<U>,
U: Hash,

Checks if an element is possibly in the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let mut filter = CuckooFilter::<String>::new(100);

filter.insert("foo");
assert!(filter.contains("foo"));

pub fn clear(&mut self)

Clears the cuckoo filter, removing all elements.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let mut filter = CuckooFilter::<String>::new(100);

filter.insert("foo");
filter.clear();

assert!(!filter.contains("foo"));

pub fn len(&self) -> usize

Returns the number of occupied entries in the cuckoo filter. It does account for items in the extra items vector.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);

assert_eq!(filter.len(), 0);

pub fn is_empty(&self) -> bool

Returns true if there are no occupied entries in the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);

assert!(filter.is_empty());

pub fn capacity(&self) -> usize

Returns the maximum capacity of the cuckoo filter. Items stay spill into the extra items vector even through the length of the cuckoo filter is less than the capacity.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);

assert_eq!(filter.capacity(), 128);

pub fn bucket_len(&self) -> usize

Returns the length of each bucket in the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);

assert_eq!(filter.bucket_len(), 32);

pub fn entries_per_index(&self) -> usize

Returns the number of entries per index in the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);

assert_eq!(filter.entries_per_index(), 4);

pub fn extra_items_len(&self) -> usize

Returns the number of items that could not be inserted into the CuckooFilter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;
use probabilistic_collections::SipHasherBuilder;

let mut filter = CuckooFilter::<String>::from_parameters_with_hashers(
    1,
    8,
    1,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

filter.insert("foo");
filter.insert("foobar");
assert_eq!(filter.extra_items_len(), 1);

pub fn is_nearly_full(&self) -> bool

Returns true if there are any items that could not be inserted into the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let mut filter = CuckooFilter::<String>::from_parameters(1, 8, 1);

filter.insert("foo");
filter.insert("foobar");
assert!(filter.is_nearly_full());

pub fn fingerprint_bit_count(&self) -> usize

Returns the number of bits in each item fingerprint.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);

assert_eq!(filter.fingerprint_bit_count(), 8);

pub fn estimated_fpp(&self) -> f64

Returns the estimated false positive probability of the cuckoo filter. This value will increase as more items are added.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let mut filter = CuckooFilter::<String>::new(100);
assert!(filter.estimated_fpp() < std::f64::EPSILON);

filter.insert("foo");
assert!(filter.estimated_fpp() > std::f64::EPSILON);
assert!(filter.estimated_fpp() < 0.01);

pub fn hashers(&self) -> &[B; 2]

Returns a reference to the cuckoo filter's hasher builders.

Examples

use probabilistic_collections::cuckoo::CuckooFilter;

let filter = CuckooFilter::<String>::new(100);
let hashers = filter.hashers();