Struct probabilistic_collections::cuckoo::ScalableCuckooFilter

pub struct ScalableCuckooFilter<T, B = SipHasherBuilder> { /* fields omitted */ }

A growable, space-efficient probabilistic data structure to test for membership in a set. Scalable cuckoo filters also provide the flexibility to remove items.

A cuckoo filter is based on cuckoo hashing and is essentially a cuckoo hash table storing each keys' fingerprint. Cuckoo filters can be highly compact and serve as an improvement over variations of tradition Bloom filters that support deletion (E.G. counting Bloom filters).

This implementation is a scalable version of a cuckoo filter inspired by ScalableBloomFilter. Currently, the scalable cuckoo filter will naively insert into the last inserted cuckoo filter despite the fact that deletions could free space in previously inserted cuckoo filters. Checking if there is space in previously inserted cuckoo filters is fairly expensive and would significantly slow down the scalable cuckoo filter.

The overall false positive probability of the scalable cuckoo filter will be initial_fpp * 1 / (1 - tightening_ratio).

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let mut filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);
assert!(!filter.contains("foo"));

filter.insert("foo");
assert!(filter.contains("foo"));

filter.remove("foo");
assert!(!filter.contains("foo"));

assert_eq!(filter.len(), 0);
assert_eq!(filter.capacity(), 128);
assert_eq!(filter.filter_count(), 1);

Methods

impl<T> ScalableCuckooFilter<T>

pub fn new(
item_count: usize,
fpp: f64,
growth_ratio: f64,
tightening_ratio: f64
) -> Self

Constructs a new, empty ScalableCuckooFilter with an estimated initial item capacity of item_count, and an initial maximum false positive probability of fpp. Every time a new cuckoo filter is added, the size will be approximately growth_ratio multiplied by the previous size, and the false positive probability will be tightening_ratio multipled by the previous false positive probability.

The length of each bucket will be rounded off to the next power of two.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

pub fn from_entries_per_index(
item_count: usize,
fpp: f64,
entries_per_index: usize,
growth_ratio: f64,
tightening_ratio: f64
) -> Self

Constructs a new, empty ScalableCuckooFilter with an estimated initial item capacity of item_count, an initial maximum false positive probability of fpp, and entries_per_index entries per index. Every time a new cuckoo filter is added, the size will be approximately growth_ratio multiplied by the previous size, and the false positive probability will be tightening_ratio multipled by the previous false positive probability.

The length of each bucket will be rounded off to the next power of two.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::from_entries_per_index(100, 0.01, 4, 2.0, 0.5);

impl<T, B> ScalableCuckooFilter<T, B> where
B: BuildHasher + Clone + Copy,

pub fn with_hashers(
item_count: usize,
fpp: f64,
growth_ratio: f64,
tightening_ratio: f64,
hash_builders: [B; 2]
) -> Self

Constructs a new, empty ScalableCuckooFilter with an estimated initial item capacity of item_count, an initial maximum false positive probability of fpp, and two hasher builders for double hashing. Every time a new cuckoo filter is added, the size will be approximately growth_ratio multiplied by the previous size, and the false positive probability will be tightening_ratio multipled by the previous false positive probability.

The length of each bucket will be rounded off to the next power of two.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;
use probabilistic_collections::SipHasherBuilder;

let filter = ScalableCuckooFilter::<String>::with_hashers(
    100,
    0.01,
    2.0,
    0.5,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

pub fn from_entries_per_index_with_hashers(
item_count: usize,
fpp: f64,
entries_per_index: usize,
growth_ratio: f64,
tightening_ratio: f64,
hash_builders: [B; 2]
) -> Self

Constructs a new, empty ScalableCuckooFilter with an estimated initial item capacity of item_count, an initial maximum false positive probability of fpp, entries_per_index entries per index, and two hasher builders for double hashing. Every time a new cuckoo filter is added, the size will be approximately growth_ratio multiplied by the previous size, and the false positive probability will be tightening_ratio multipled by the previous false positive probability.

The length of each bucket will be rounded off to the next power of two.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;
use probabilistic_collections::SipHasherBuilder;

let filter = ScalableCuckooFilter::<String>::from_entries_per_index_with_hashers(
    100,
    0.01,
    4,
    2.0,
    0.5,
    [SipHasherBuilder::from_seed(0, 0), SipHasherBuilder::from_seed(1, 1)],
);

pub fn insert<U: ?Sized>(&mut self, item: &U) where
T: Borrow<U>,
U: Hash,

Inserts an element into the scalable cuckoo filter.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let mut filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

filter.insert("foo");

pub fn contains<U: ?Sized>(&mut self, item: &U) -> bool where
T: Borrow<U>,
U: Hash,

Checks if an element is possibly in the scalable cuckoo filter.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let mut filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

assert!(!filter.contains("foo"));
filter.insert("foo");
assert!(filter.contains("foo"));

pub fn remove<U: ?Sized>(&mut self, item: &U) where
T: Borrow<U>,
U: Hash,

Removes an element from the scalable cuckoo filter.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let mut filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);
assert!(!filter.contains("foo"));

filter.insert("foo");
assert!(filter.contains("foo"));

filter.remove("foo");
assert!(!filter.contains("foo"));

pub fn len(&self) -> usize

Returns the number of occupied entries in the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

assert_eq!(filter.len(), 0);

pub fn is_empty(&self) -> bool

Returns true if there are no occupied entries in the cuckoo filter.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

assert!(filter.is_empty());

pub fn capacity(&self) -> usize

Returns the maximum capacity of the cuckoo filter. The scalable filter may grow even through the length of the cuckoo filter is less than the capacity.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

assert_eq!(filter.capacity(), 128);

pub fn entries_per_index(&self) -> usize

Returns the number of entries per index in each filter.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

assert_eq!(filter.entries_per_index(), 4);

pub fn filter_count(&self) -> usize

Returns the number of cuckoo filters used by the scalable cuckoo filter.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

assert_eq!(filter.filter_count(), 1);

pub fn clear(&mut self)

Clears the scalable cuckoo filter, removing all elements.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let mut filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);

filter.insert("foo");
filter.clear();

assert!(!filter.contains("foo"));

pub fn estimated_fpp(&self) -> f64

Returns the estimated false positive probability of the scalable cuckoo filter. This value will increase as more items are added.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let mut filter = ScalableCuckooFilter::::new(100, 0.01, 2.0, 0.5); assert!(filter.estimated_fpp() < std::f64::EPSILON);

filter.insert("foo"); assert!(filter.estimated_fpp() > std::f64::EPSILON); assert!(filter.estimated_fpp() < 0.01);

pub fn hashers(&self) -> &[B; 2]

Returns a reference to the scalable cuckoo filter's hasher builders.

Examples

use probabilistic_collections::cuckoo::ScalableCuckooFilter;

let filter = ScalableCuckooFilter::<String>::new(100, 0.01, 2.0, 0.5);
let hashers = filter.hashers();